Renamed Async.Future to ScopedFuture
