+++ core/src/main/scala/fs2/task.scala
@@ -1,8 +1,8 @@
 package fs2
 
-import fs2.internal.{Actor, Future, LinkedMap}
-import fs2.util.Effect
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
+import fs2.internal.{Actor,Future,LinkedMap}
+import fs2.util.{Effect}
+import java.util.concurrent.atomic.{AtomicBoolean,AtomicReference}
 
 import scala.concurrent.duration._
